Single Post Page and Delete Button
Clicking on a post links the postId and navigates the page to that postId specifically. Also made the delete button function from the home page and the single post page. Made it so that deleting a post will send the user back to the homepage and the post will be gone, the user no longer has to refresh the page in order to see the changes in the front end